(CONT-803) Reubocop Auto Fixes 21-26
- Style/IfUnlessModifier
- Style/NegatedIfElseCondition
- Style/RedundantAssignment
- Style/RescueStandardError
- Style/SoleNestedConditional
- Style/TrailingCommaInHashLiteral

+++ lib/puppet/type/vcsrepo.rb
@@ -125,19 +125,13 @@ Puppet::Type.newtype(:vcsrepo) do
     end
 
     newvalue :absent do
-      if provider.exists?
-        provider.destroy
-      end
+      provider.destroy if provider.exists?
     end
 
     newvalue :latest, required_features: [:reference_tracking] do
       if provider.exists? && !@resource.value(:force)
-        if provider.class.feature?(:bare_repositories) && provider.bare_exists?
-          provider.convert_bare_to_working_copy
-        end
-        if provider.respond_to?(:update_references)
-          provider.update_references
-        end
+        provider.convert_bare_to_working_copy if provider.class.feature?(:bare_repositories) && provider.bare_exists?
+        provider.update_references if provider.respond_to?(:update_references)
         reference = if provider.respond_to?(:latest?)
                       provider.latest || provider.revision
                     else
@@ -174,9 +168,7 @@ Puppet::Type.newtype(:vcsrepo) do
     isnamevar
     validate do |value|
       path = Pathname.new(value)
-      unless path.absolute?
-        raise ArgumentError, "Path must be absolute: #{path}"
-      end
+      raise ArgumentError, "Path must be absolute: #{path}" unless path.absolute?
     end
   end
 
@@ -250,9 +242,7 @@ Puppet::Type.newtype(:vcsrepo) do
   newparam :compression, required_features: [:gzip_compression] do
     desc 'Compression level'
     validate do |amount|
-      unless Integer(amount).between?(0, 6)
-        raise ArgumentError, "Unsupported compression level: #{amount} (expected 0-6)"
-      end
+      raise ArgumentError, "Unsupported compression level: #{amount} (expected 0-6)" unless Integer(amount).between?(0, 6)
     end
   end
 
@@ -348,14 +338,10 @@ Puppet::Type.newtype(:vcsrepo) do
     validate do |value|
       if value.is_a?(Hash)
         value.each do |k, v|
-          unless v.match?(regex)
-            raise ArgumentError, "HTTP Proxy for #{k} must be an HTTP/HTTPS URL: #{v}"
-          end
+          raise ArgumentError, "HTTP Proxy for #{k} must be an HTTP/HTTPS URL: #{v}" unless v.match?(regex)
         end
       else
-        unless value.match?(regex)
-          raise ArgumentError, "HTTP Proxy must be an HTTP/HTTPS URL: #{value}"
-        end
+        raise ArgumentError, "HTTP Proxy must be an HTTP/HTTPS URL: #{value}" unless value.match?(regex)
       end
       # Validation passed.
       super(value)
